engine_traits: rename cfdefs to cf_defs
