+++ components/engine_traits/src/lib.rs
@@ -241,8 +241,8 @@ pub use crate::peekable::*;
 // These modules contain support code that does not need to be implemented by
 // engines.
 
-mod cfdefs;
-pub use crate::cfdefs::*;
+mod cf_defs;
+pub use crate::cf_defs::*;
 mod engines;
 pub use engines::*;
 mod errors;
